Finished Analyze Selection UX / UI
Some issues found:

triple-clicking a line in vscode, will make it select with a new line, which is a bit trouble some.

Currently the code works so that selecting, like clicking and hovering over the text works.

Currently it just reverses the current code, and should be replaced with an actually http service call.